create file only once in parallel
