+++ src/Console/Command/ProcessCommand.php
@@ -17,7 +17,6 @@ use Rector\Core\Exception\ShouldNotHappenException;
 use Rector\Core\Reporting\MissingRectorRulesReporter;
 use Rector\Core\StaticReflection\DynamicSourceLocatorDecorator;
 use Rector\Core\Validation\EmptyConfigurableRectorChecker;
-use Rector\Core\ValueObject\Application\File;
 use Rector\Core\ValueObject\Configuration;
 use Rector\Core\ValueObject\ProcessResult;
 use Rector\Core\ValueObjectFactory\Application\FileFactory;
@@ -93,18 +92,12 @@ final class ProcessCommand extends AbstractProcessCommand
         // 2. inform user about registering configurable rule without configuration
         $this->emptyConfigurableRectorChecker->check();
 
-        // 3. collect all files from files+dirs provided paths
-        $files = $this->fileFactory->createFromPaths($paths, $configuration);
-
-        // 4. PHPStan has to know about all files too
-        $this->configurePHPStanNodeScopeResolver($files);
-
         // MAIN PHASE
-        // 5. run Rector
-        $systemErrorsAndFileDiffs = $this->applicationFileProcessor->run($files, $configuration, $input);
+        // 3. run Rector
+        $systemErrorsAndFileDiffs = $this->applicationFileProcessor->run($configuration, $input);
 
         // REPORTING PHASE
-        // 6. reporting phase
+        // 4. reporting phase
         // report diffs and errors
         $outputFormat = $configuration->getOutputFormat();
         $outputFormatter = $this->outputFormatterCollector->getByName($outputFormat);
@@ -157,34 +150,5 @@ final class ProcessCommand extends AbstractProcessCommand
         }
 
         return $processResult->getFileDiffs() === [] ? Command::SUCCESS : Command::FAILURE;
-    }
-
-    /**
-     * @param File[] $files
-     */
-    private function configurePHPStanNodeScopeResolver(array $files): void
-    {
-        $filePaths = $this->resolvePhpFilePaths($files);
-        $this->nodeScopeResolver->setAnalysedFiles($filePaths);
-    }
-
-    /**
-     * @param File[] $files
-     * @return string[]
-     */
-    private function resolvePhpFilePaths(array $files): array
-    {
-        $filePaths = [];
-
-        foreach ($files as $file) {
-            $smartFileInfo = $file->getSmartFileInfo();
-            $pathName = $smartFileInfo->getPathname();
-
-            if (\str_ends_with($pathName, '.php')) {
-                $filePaths[] = $pathName;
-            }
-        }
-
-        return $filePaths;
     }
 }
